pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "capybara-docsaid"
dynamic = ["version"]
description = "An Image Processing and Deep Learning Toolkit."
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "Apache License 2.0"}
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "License :: OSI Approved :: Apache Software License",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
  "dacite",
  "requests",
  "numpy",
  "pdf2image",
  "ujson",
  "pyyaml",
  "tqdm",
  "pybase64",
  "PyTurboJPEG",
  "dill",
  "natsort",
  "shapely",
  "piexif",
  "opencv-python>=4.12.0.88",
  "beautifulsoup4",
  "pillow-heif"
]

[project.optional-dependencies]
# Inference backends (opt-in; keep `pip install .` minimal)
onnxruntime = [
  "onnxruntime>=1.22.0,<2",
  "onnx>=1.18.0",
  "onnxslim>=0.1.0",
]
onnxruntime-gpu = [
  "onnxruntime-gpu>=1.22.0,<2",
  "onnx>=1.18.0",
  "onnxslim>=0.1.0",
]
openvino = [
  "openvino>=2024.0.0",
]
torchscript = [
  "torch>=2.0",
]
all = [
  "onnxruntime>=1.22.0,<2",
  "onnx>=1.18.0",
  "onnxslim>=0.1.0",
  "openvino>=2024.0.0",
  "torch>=2.0",
]

# Feature extras (non-core runtime)
ipcam = [
  "flask>=2.0",
]
system = [
  "psutil",
]
visualization = [
  "matplotlib",
  "pillow",
]

[project.urls]
Homepage = "https://docsaid.org/en/docs/capybara/"
Repository = "https://github.com/DocsaidLab/Capybara"
Issues = "https://github.com/DocsaidLab/Capybara/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["capybara*"]
exclude = [
  "demo",
  "tests",
  "docs",
  ".github",
  "docker",
  "wheelhouse"
]

[tool.setuptools.dynamic]
version = { attr = "capybara.__version__" }

[tool.pytest.ini_options]
pythonpath = [
  ".",
]
testpaths = [
  "tests",
]
addopts = "--ignore=tmp"

[tool.ruff]
target-version = "py310"
line-length = 80
extend-exclude = [
  ".git",
  ".mypy_cache",
  ".pytest_cache",
  "__pycache__",
  ".venv",
  "venv",
  "build",
  "dist",
  ".tox",
  ".eggs",
  "*.egg-info",
]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "W",
  "B",
  "UP",
  "N",
  "I",
  "C4",
  "SIM",
  "RUF",
]
ignore = [
  "E203",
  "E501",
]

[tool.ruff.lint.isort]
known-first-party = ["capybara", "tests"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
